feat(scenarios): unify all scenario runs
